Move some core tests into render and parts tests
