+++ test/index.html
@@ -7,13 +7,15 @@
   <body>
     <script type="module" src="./core_test.js"></script>
     <script type="module" src="./lit-html_test.js"></script>
-    <script type="module" src="./lib/repeat_test.js"></script>
-    <script type="module" src="./lib/until_test.js"></script>
+    <script type="module" src="./lib/parts_test.js"></script>
+    <script type="module" src="./lib/render_test.js"></script>
     <script type="module" src="./lib/lit-extended_test.js"></script>
-    <script type="module" src="./lib/unsafe-html_test.js"></script>
-    <script type="module" src="./lib/async-append_test.js"></script>
-    <script type="module" src="./lib/async-replace_test.js"></script>
     <script type="module" src="./lib/modify-template_test.js"></script>
+    <script type="module" src="./directives/repeat_test.js"></script>
+    <script type="module" src="./directives/until_test.js"></script>
+    <script type="module" src="./directives/unsafe-html_test.js"></script>
+    <script type="module" src="./directives/async-append_test.js"></script>
+    <script type="module" src="./directives/async-replace_test.js"></script>
     <script type="module">
       WCT.loadSuites([
         'shady.html',
